Reorganize dcd fils
